+++ src/db/migrations/0011_panoramic_hobgoblin.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `playlists` ADD `source` text;--> statement-breakpoint
+UPDATE `playlists` SET `source` = CASE WHEN `spotify_id` IS NOT NULL THEN 'spotify' ELSE 'local' END WHERE `source` IS NULL;